requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
openpyxl>=3.1.0
rapidfuzz>=3.13.0